refactor: themed icons
