+++ site/src/components/home/sections/latestTransfers.js
@@ -7,8 +7,6 @@ import {
 } from "../../../styles/text";
 import { withLoading } from "../../../HOC/withLoading";
 import Loading from "../../loadings/loading";
-import { ReactComponent as Transfer } from "./transfer.svg";
-import { ReactComponent as TransferRight } from "./transferRight.svg";
 import { ReactComponent as CheckIcon } from "../../icons/check.svg";
 import { ReactComponent as TimerIcon } from "../../icons/timer.svg";
 import { timeDuration } from "../../../utils/viewFuncs/time";
@@ -19,28 +17,8 @@ import { useSelector } from "react-redux";
 import { chainSettingSelector } from "../../../store/reducers/settingSlice";
 import { OnlyDesktop } from "../../screen/onlyDesktop";
 import Tooltip from "../../tooltip";
-
-const TransferIcon = styled(Transfer)`
-  path {
-    stroke: ${({ theme }) => theme.fontTertiary};
-  }
-
-  linearGradient {
-    stop:first-child {
-      stop-color: ${({ theme }) => theme.fontPrimary};
-    }
-
-    stop:last-child {
-      stop-color: ${({ theme }) => theme.fontPrimary};
-    }
-  }
-`;
-
-const TransferRightIcon = styled(TransferRight)`
-  path {
-    stroke: ${({ theme }) => theme.fontTertiary};
-  }
-`;
+import TransferIcon from "../../icons/transferIcon";
+import TransferRightIcon from "../../icons/transferRight";
 
 const Rows = styled.ul`
   margin: 0;
